refactor : 즐겨찾기 관련 클래스들 favorite 패키지로 재이동
