+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,12 +1,13 @@
-package nextstep.subway.station.application;
+package nextstep.subway.favorite.application;
 
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
-import nextstep.subway.station.domain.Favorite;
-import nextstep.subway.station.domain.FavoriteRepository;
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.favorite.domain.Favorite;
+import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.dto.FavoriteRequest;
+import nextstep.subway.favorite.dto.FavoriteRequest;
 import org.springframework.stereotype.Service;
 
 @Service
