+++ src/main/java/nextstep/subway/favorite/domain/Favorite.java
@@ -1,7 +1,8 @@
-package nextstep.subway.station.domain;
+package nextstep.subway.favorite.domain;
 
 import nextstep.subway.BaseEntity;
 import nextstep.subway.member.domain.Member;
+import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
