+++ src/main/java/nextstep/subway/favorite/domain/FavoriteRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station.domain;
+package nextstep.subway.favorite.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
