+++ src/main/java/nextstep/subway/member/domain/Member.java
@@ -2,7 +2,7 @@ package nextstep.subway.member.domain;
 
 import nextstep.subway.BaseEntity;
 import nextstep.subway.auth.application.AuthorizationException;
-import nextstep.subway.station.domain.Favorite;
+import nextstep.subway.favorite.domain.Favorite;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.persistence.*;
