+++ src/test/java/nextstep/subway/favorite/FavoriteAcceptanceTest.java
@@ -6,7 +6,7 @@ import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.station.dto.FavoriteRequest;
+import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
